requirements.txt
 annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.0
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.7
packaging==24.1
passlib==1.7.4
pendulum==3.0.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
pyasn1==0.6.0
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
redis==4.6.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.32
starlette==0.37.2
time-machine==2.15.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.5
watchfiles==0.23.0
websockets==12.0
win32-setctime==1.1.0
WTForms==3.1.2
yarl==1.9.4
shapely==2.0.6
matplotlib==3.9.2